chore: add affecting filter message in the report header
