+++ src/Components/SmartComponents/CVEs/CVEsTableToolbar.js
@@ -197,7 +197,10 @@ const CVEsTableToolbarWithContext = ({ context, canEditStatusOrBusinessRisk, can
                 <DownloadCVEsReport
                     showButton={false}
                     params={params}
-                    filters={buildActiveFilters(params, isEdgeParityEnabled)}
+                    filters={{
+                        ...buildActiveFilters(params, isEdgeParityEnabled),
+                        affecting: { value: isEdgeParityEnabled ? 'rpmdnf,edge' : 'true' }
+                    }}
                     onSuccess={() => setExportPDF(false)}
                 />
             }
