+++ src/Components/SmartComponents/Reports/Common/firstPagePDF/helpers.js
@@ -1,0 +1,25 @@
+import messages from '../../../../../Messages';
+
+export const getAffectingFilterTexts = (systemCount, conventionalCount, edgeCount) => ({
+    false: `not affecting any ${systemCount}`,
+    true: `impacting at least one of your ${systemCount} analyzed RHEL systems.`,
+    rpmdnf: `impacting at least one of your analyzed ${systemCount} RHEL conventional systems.`,
+    edge: 'impacting at least one of your analyzed ${systemCount} RHEL immutable systems.',
+    'rpmdnf,edge': `impacting at least one of your analyzed ${systemCount} RHEL systems, 
+    consisting of ${conventionalCount} conventional and ${edgeCount} immutable.`
+});
+
+export const cveFilterStringSegments = [
+    { name: 'impact', msg: messages.CVEsPDFreportFiltersSeverity, lowercase: true },
+    { name: 'business_risk_id', msg: messages.CVEsPDFreportFiltersBusinessRisk, lowercase: true },
+    { name: 'status_id', msg: messages.CVEsPDFreportFiltersStatus, lowercase: true },
+    { name: 'rule_presence', msg: messages.CVEsPDFreportFiltersDefinedValue, lowercase: true },
+    { name: 'known_exploit', msg: messages.CVEsPDFreportFiltersDefinedValue, lowercase: true },
+    { name: 'advisory_available', msg: messages.CVEsPDFreportFiltersAdvisoryAvailable, lowercase: true },
+    { name: 'rhel_version', msg: messages.CVEsPDFreportFiltersOsVersion }
+];
+
+export const systemsFilterStringSegments = [
+    { name: 'filter', msg: messages.systemsPDFreportFiltersSearchTerm, lowercase: true },
+    { name: 'rhel_version', msg: messages.systemsPDFreportFiltersOsVersion }
+];
