+++ src/Components/SmartComponents/Reports/Common/firstPagePDF/index.js
@@ -1,0 +1,1 @@
+export { default } from './firstPagePDF';
