+++ src/Components/SmartComponents/Reports/Common/footerPDF/FooterPDF.js
@@ -1,12 +1,11 @@
 import React from 'react';
 import propTypes from 'prop-types';
 import { View, Text } from '@react-pdf/renderer';
-import styles from './styles';
-import messages from '../../../../Messages';
-import { intlFormatWithBold } from '../../../../Helpers/ReportsHelper';
+import styles from '../styles';
+import messages from '../../../../../Messages';
+import { intlFormatWithBold } from '../../../../../Helpers/ReportsHelper';
 
 const FooterPDF = ({ hasRule, hasExploit }) => {
-
     return (
         <View>
             { hasExploit && (
