+++ src/Components/SmartComponents/Reports/Common/footerPDF/FooterPDF.test.js
