+++ src/Components/SmartComponents/Reports/Common/footerPDF/index.js
@@ -1,0 +1,1 @@
+export { default } from './FooterPDF';
