+++ src/Components/SmartComponents/Reports/Common/tablePage/index.js
@@ -1,0 +1,1 @@
+export { default } from './tablePage';
