+++ src/Components/SmartComponents/Reports/Common/tablePage/tablePage.js
@@ -3,11 +3,11 @@ import PropTypes from 'prop-types';
 import { View, Text, Link } from '@react-pdf/renderer';
 import { Table } from '@redhat-cloud-services/frontend-components-pdf-generator';
 import { processDate } from '@redhat-cloud-services/frontend-components-utilities/helpers';
-import { CVES_PATH } from '../../../../Helpers/constants';
-import messages from '../../../../Messages';
-import styles from './styles';
-import { formatDate } from '../../../../Helpers/MiscHelper';
-import CVElabels from '../Common/CVElabels';
+import { CVES_PATH } from '../../../../../Helpers/constants';
+import messages from '../../../../../Messages';
+import styles from '../styles';
+import { formatDate } from '../../../../../Helpers/MiscHelper';
+import CVElabels from '../CVElabels';
 
 const tablePage = ({ data, page, intl, header, type, isReportDynamic = false }) => {
     const POSSIBLE_HEADERS = [
