+++ src/Components/SmartComponents/Reports/Common/tablePage/tablePage.test.js
@@ -1,10 +1,10 @@
 import TablePage from './tablePage';
 import toJson from 'enzyme-to-json';
-import { mountWithIntl } from '../../../../Helpers/MiscHelper';
+import { mountWithIntl } from '../../../../../Helpers/MiscHelper';
 import { BrowserRouter as Router } from 'react-router-dom';
 import configureStore from 'redux-mock-store';
 import { Provider, useSelector } from 'react-redux';
-import { intl } from '../../../../Utilities/IntlProvider';
+import { intl } from '../../../../../Utilities/IntlProvider';
 
 const mockStore = configureStore([store => next => action => {}]);
 let store = mockStore({ });
