+++ src/Components/SmartComponents/Reports/DownloadCVEsReport.js
@@ -9,7 +9,7 @@ import { firstPagePDFLength } from '../../../Helpers/ReportsHelper';
 import messages from '../../../Messages';
 import firstPagePDF from './Common/firstPagePDF';
 import tablePage from './Common/tablePage';
-import FooterPDF from './Common/FooterPDF';
+import FooterPDF from './Common/footerPDF';
 import DownloadReport from '../../../Helpers/DownloadReport';
 import {
     STATUS_OPTIONS,
@@ -59,7 +59,6 @@ const DownloadCVEsReport = ({
                 ...params,
                 ...DownloadReport.defaultParams,
                 sort: params.sort || '-cvss_score',
-                affecting: 'true',
                 ...isReportDynamic
                     ? { advanced_report: 'true' }
                     : { report: 'true' }
