+++ src/Helpers/ReportsHelper.js
@@ -92,8 +92,8 @@ export const buildFilters = (filterData, isEdgeEnabled) => {
             case 'affecting':
                 newValues[key].values = value.map(
                     checkedValue =>
-                        AFFECTING_FILTER_OPTIONS.find(({ value }) => value === checkedValue).label
-                ).join(', ');
+                        AFFECTING_FILTER_OPTIONS.find(({ value }) => value === checkedValue).value
+                ).join(',');
                 break;
             case 'tags':
                 newValues[key].values = value;
