+++ src/Messages.js
@@ -459,7 +459,7 @@ export default defineMessages({
     CVEsPDFreportSubHeader: {
         id: 'CVEsReport.subheader',
         description: 'CVEs PDF report sub header',
-        defaultMessage: 'The vulnerability service identified <b>{cve_count, plural, one {# CVE} other {# CVEs}}</b> within this criteria that impact at least one of your <b>{system_count} analyzed RHEL systems</b>.\n'
+        defaultMessage: 'The vulnerability service identified <b>{cve_count, plural, one {# CVE} other {# CVEs}}</b> within this criteria, {affectingSystemsMessage}\n'
     },
     CVEsPDFreportSubHeaderExploit: {
         id: 'CVEsReport.subheaderExploit',
